requirements.txt
fastapi
uvicorn[standard]
transformers
Pillow
python-multipart
diffusers
invisible_watermark
accelerate
safetensors
datasets
scipy
streamlit